Enable "poll" command only for moderators in installation instructions
The "poll" command makes possible to create new polls as well as close
the current poll, no matter which user created it. Therefore, it is best
suited as a command available only to moderators of the conversation
(but it can be made available to users too if needed; this only changes
the recommended way to configure it).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>